requirements.txt
Django>=5.0,<6
channels>=4.0
daphne>=4.0
channels-redis>=4.0
python-decouple>=3.8
psycopg2-binary>=2.9
pymongo>=4.6
Pillow>=10.0
python-docx>=1.1
django-crispy-forms>=2.0
crispy-bootstrap5>=2023.0
django-filter>=23.0
openpyxl>=3.0
reportlab>=4.0

# REST API
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# AI-поиск
langchain>=0.3
langchain-groq>=0.2
requests>=2.31
